+++ src/ggrc/assets/javascripts/components/people/editable-people-group.js
@@ -5,10 +5,11 @@
 
 import './people-group-modal';
 import './editable-people-group-header';
+import peopleGroupVM from '../view-models/people-group-vm';
 
 const SHOW_MODAL_LIMIT = 4;
 
-var viewModel = GGRC.VM.DeletablePeopleGroup.extend({
+var viewModel = peopleGroupVM.extend({
   title: '@',
   canEdit: {},
   showPeopleGroupModal: false,
